Don't read and write grades.json on every request
CO-569

Instead, use a global grades object which is read at application startup and
written at shutdown.

Fixes a bunch of horrible race conditions.

The original reason for reading the json file on every request was so
users could be added easily during a demo, but that isn't necessary now
that we generate random grades for unknown users.